refactored to move all components form home page to components/homepage
