+++ src/components/homepage/car_card.jsx
