+++ src/components/homepage/card_carousel.jsx
